% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{api_ccodwg}
\alias{api_ccodwg}
\title{Generate API call for the COVID-19 Canada Open Data Working Group dataset}
\usage{
api_ccodwg(type, args_list)
}
\arguments{
\item{type}{One of "timeseries" (time series data), "summary" (summary data),
"individual" (individual-level data),  or "other" (supplementary files).}

\item{args_list}{A list of API arguments for the selected data type.}
}
\value{
A character string representing the API call with the specified
arguments.
}
\description{
This function is used within dl_ccodwg() to generate the API call with the
specified arguments.
}
