% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_ccodwg.R
\name{ccodwg_update_date}
\alias{ccodwg_update_date}
\title{Get date the COVID-19 Canada Open Data Working Group dataset was last updated}
\usage{
ccodwg_update_date()
}
\description{
This is a simple convenience function to return the date portion of a
"version" call to the API.
}
\examples{
# get date the CCODWG dataset was last updated
ccodwg_update_date()
}
