% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_curl_handle}
\alias{create_curl_handle}
\title{Create curl handle with relevant options set}
\usage{
create_curl_handle(d)
}
\arguments{
\item{d}{Information on dataset from \code{\link[Covid19CanadaData]{get_uuid}}.}
}
\value{
A `curl` handle with relevant options set.
}
\description{
Helper function for \code{\link[Covid19CanadaData]{dl_dataset}}.
}
