% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_ccodwg.R
\name{dl_ccodwg}
\alias{dl_ccodwg}
\title{Download data from the COVID-19 Canada Open Data Working Group dataset}
\usage{
dl_ccodwg(
  type = c("timeseries", "individual", "summary", "other", "version"),
  stat = c("cases", "mortality", "recovered", "testing", "active", "avaccine",
    "dvaccine", "cvaccine"),
  loc = c("default", "canada", "prov", "hr"),
  date = NA,
  after = NA,
  before = NA,
  ymd = "true",
  missing = "na",
  extra = "true",
  verbose = FALSE,
  file = NULL
)
}
\arguments{
\item{type}{One of "timeseries" (time series data), "summary" (summary data),
"individual" (individual-level data), "other" (supplementary files) or
"version" (date and time dataset was last updated).}

\item{stat}{Which statistic to return. One of "cases", "mortality",
"recovered", "testing", "active", "avaccine", "dvaccine" or "cavvacine".
Note that not all datasets are available at all geographic levels.}

\item{loc}{The geographic level of the data. One of "default", "canada",
"prov", "hr", a two-letter province code or a health region code. Not all
geographic levels are available for all datasets. See API documentation for
all possible values. If NA, the default value for the specified dataset will
be used.}

\item{date}{A character string specifying the date of data to return.
Use either YYYY-MM-DD or DD-MM-YYYY format.}

\item{after}{A character string specifying that data from this date or later
should be returned. Use either YYYY-MM-DD or DD-MM-YYYY format.}

\item{before}{A character string specifying that data from this date or
earlier should be returned. Use either YYYY-MM-DD or DD-MM-YYYY format.}

\item{ymd}{One of "true" or "false". Should dates be returned in YYYY-MM-DD
format? Default = "true". Otherwise, dates are returned in DD-MM-YYYY format.}

\item{missing}{How should missing values be returned? One of "null", "na",
"nan" or "empty". Default = "na". This reads into R as NA.}

\item{extra}{One of "true" or "false". For individual-level data, should
abbreviated columns be joined? (E.g., "case_source" for individual-level
case data) Default = "true".}

\item{verbose}{Logical. Print debug messages?}

\item{file}{A character string specifying the location to write the specified
dataset as a CSV file (NULL by default, resulting in the dataset being
returned as a data frame).}
}
\value{
The specified dataset either as a data frame in R (the default) or
written to a CSV file by (by specifying the `file` argument). If
type = "version", the date and time the dataset were last updated is returned
as a character string.
}
\description{
Download data from the COVID-19 Canada Open Data Working Group dataset via
the JSON API. Pre-processing of the data may be done by specifying the
relevant arguments in the function which are then passed to the API call.
}
\details{
Full documentation of the API is available here: https://opencovid.ca/api/

Data can either be imported into R (the default) or written to a CSV file by
specifying the `file` argument.
}
\examples{
# get case time series for Toronto during the first half of March 2020
dl_ccodwg("timeseries", "cases", loc = 3595, after = "2020-03-01", before = "2020-03-15")

# get most recent Canada-wide summary
dl_ccodwg("summary", loc = "canada")

# get date the dataset was last updated
as.Date(dl_ccodwg("version"))
}
