% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_ccodwg.R
\name{dl_dataset}
\alias{dl_dataset}
\title{Download current version of a dataset catalogued in Covid19CanadaArchive}
\usage{
dl_dataset(uuid, file = NULL)
}
\arguments{
\item{uuid}{The UUID of the dataset from datasets.json.}

\item{file}{A character string specifying the location to write the specified
dataset as a file (NULL by default, resulting in the dataset being returned
as a data frame).}
}
\value{
The specified dataset either as a data frame in R (the default) or
written to a file by (by specifying the argument `file`).
}
\description{
Download the current version of an active dataset listed in datasets.json of
Covid19CanadaArchive (https://github.com/ccodwg/Covid19CanadaArchive/blob/master/data/datasets.json).
Data can either be imported into R (the default) or written to a file by
specifying the `file` argument. Currently, only CSV datasets are supported.
}
\examples{
\dontrun{
# get PHAC epidemiology update CSV
dl_dataset("f7db31d0-6504-4a55-86f7-608664517bdb")

# get Saskatchewan total cases CSV
dl_dataset("61cfdd06-7749-4ae6-9975-d8b4f10d5651")
}
}
