% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dl_dataset_dyn_url}
\alias{dl_dataset_dyn_url}
\title{Download current version of a dataset catalogued in Covid19CanadaArchive: Get dynamic URL}
\usage{
dl_dataset_dyn_url(uuid)
}
\arguments{
\item{uuid}{The UUID of the dataset from datasets.json.}
}
\value{
The current URL of the specified dataset.
}
\description{
Helper function for \code{\link[Covid19CanadaData]{dl_dataset}}:
data-specific code to retrieve current URL of a dataset with dynamic URLs.
Replicates code included in the "url_fun_r" field of datasets.json.
This code is intentionally written to fit on a single line.
}
