% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dl_dataset_e}
\alias{dl_dataset_e}
\alias{e_uuid}
\alias{e_val}
\alias{e_fmt}
\title{Error functions for dl_dataset}
\usage{
e_uuid()

e_val()

e_fmt()
}
\description{
Error functions for \code{\link{dl_dataset}}
}
