% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{ds_env}
\alias{ds_env}
\title{Persistent storage for datasets.json}
\format{
An object of class \code{environment} of length 0.
}
\usage{
ds_env
}
\description{
Persistent storage for datasets.json
}
\keyword{datasets}
