% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_dataset_arg}
\alias{get_dataset_arg}
\title{Get argument of dataset by UUID}
\usage{
get_dataset_arg(uuid, arg)
}
\arguments{
\item{uuid}{The UUID of the dataset from datasets.json.}

\item{arg}{The argument to return.}
}
\value{
The named arg of the specified dataset.
}
\description{
Get argument of dataset by UUID
}
