% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_dataset_url}
\alias{get_dataset_url}
\title{Get URL of dataset by UUID}
\usage{
get_dataset_url(uuid)
}
\arguments{
\item{uuid}{The UUID of the dataset from datasets.json.}
}
\value{
The URL of the specified dataset.
}
\description{
If the URL is dynamic, the current URL will be calculated and returned.
}
