% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_prov_names}
\alias{get_prov_names}
\title{Get province name abbreviations}
\usage{
get_prov_names()
}
\value{
A character vector of province name abbreviations.
}
\description{
Get province name abbreviations
}
