% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_sheet_names}
\alias{get_sheet_names}
\title{Get sheet names of an Excel file}
\usage{
get_sheet_names(file)
}
\arguments{
\item{file}{The location of the file to read (may also be a URL).}
}
\value{
The sheet names of the Excel file as a character vector.
}
\description{
Get sheet names of an Excel file
}
