% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_uuid}
\alias{get_uuid}
\title{Get information on a specific dataset by UUID}
\usage{
get_uuid(uuid)
}
\arguments{
\item{uuid}{The UUID of the dataset from datasets.json.}
}
\value{
A named list of available information on the dataset.
}
\description{
Get information on a specific dataset by UUID
}
