% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{index_archive}
\alias{index_archive}
\title{Call file archive index}
\usage{
index_archive(
  uuid,
  date = NULL,
  after = NULL,
  before = NULL,
  remove_duplicates = TRUE
)
}
\arguments{
\item{uuid}{The UUID of the dataset from datasets.json.}

\item{date}{A character string in YYYY-MM-DD format specifying the date of
data to return. Also accepts "latest" (the default), "first" and "all".}

\item{after}{A character string in YYYY-MM-DD format specifying that data
from this date or later should be returned. Ignored if `date` is defined.}

\item{before}{A character string in YYYY-MM-DD format specifying that data
from this date or earlier should be returned. Ignored if `date` is defined.}

\item{remove_duplicates}{Remove duplicate files from the sample after date
filtering? Defaults to `TRUE`. Ignored if a single date is requested.}
}
\value{
Archive file index matching specified UUID and date filters,
as a data frame.
}
\description{
Call file archive index
}
