% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nt.R
\name{process_nt}
\alias{process_nt}
\title{Functions to process datasets: NT}
\usage{
process_nt(uuid, val, fmt, ds, date_current, testing_type)
}
\arguments{
\item{uuid}{The UUID of the dataset from datasets.json.}

\item{val}{The desired value to be extracted from the dataset. Valid
inputs, include "cases", "mortality", "testing", etc. See details.}

\item{fmt}{The format of the output data.}

\item{ds}{The downloaded dataset in the format returned by `dl_dataset`.}
}
\value{
A standardized version of the dataset for the desired value.
}
\description{
Process a dataset listed in datasets.json of Covid19CanadaArchive (https://github.com/ccodwg/Covid19CanadaArchive/blob/master/datasets.json)
into a standardized format for a specific value provided by that dataset
(e.g., cases, deaths, hospitalizations, etc.). If `ds` is provided, the
function will read the dataset from a named list indexed by UUID. Otherwise,
the active version of the dataset will be read using \code{\link{dl_dataset}}
(if available).
}
\details{
The currently supported values are the following:
\itemize{
 \item cases
 \item mortality
 \item recovered
 \item testing
 \item vaccine_distribution
 \item vaccine_administration
 \item vaccine_completion
}
}
\keyword{internal}
