% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_dataset}
\alias{read_dataset}
\title{Helper function for \code{\link[Covid19CanadaData]{dl_dataset}} and
\code{\link[Covid19CanadaData]{dl_archive}}: determine file extension and
fetch additional parameters (if necessary) in order to read the file into R.}
\usage{
read_dataset(url, d, sep = NULL, sheet = NULL, port = NULL, host = NULL)
}
\arguments{
\item{url}{The URL of the dataset.}

\item{d}{Information on dataset from \code{\link[Covid19CanadaData]{get_uuid}}.}

\item{sep}{See the parameter in \code{\link[Covid19CanadaData]{dl_dataset}}.}

\item{sheet}{See the parameter in \code{\link[Covid19CanadaData]{dl_dataset}}.}

\item{port}{See the parameter in \code{\link[Covid19CanadaData]{dl_dataset}}.}

\item{host}{See the parameter in \code{\link[Covid19CanadaData]{dl_dataset}}.}
}
\value{
The specified dataset as an R object.
}
\description{
Helper function for \code{\link[Covid19CanadaData]{dl_dataset}} and
\code{\link[Covid19CanadaData]{dl_archive}}: determine file extension and
fetch additional parameters (if necessary) in order to read the file into R.
}
