% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableau.R
\name{tableau}
\alias{tableau}
\title{Functions for scraping Tableau dashboards}
\description{
Functions for scraping Tableau dashboards that return their content as images
rather than HTML. Raw data may be gleaned through other methods, such as
parsing HTML tooltips.
}
