% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableau.R
\name{tableau_session_id}
\alias{tableau_session_id}
\title{Get Tableau session ID}
\usage{
tableau_session_id(url, req_root)
}
\arguments{
\item{url}{The URL of the Tableau dashboard.}

\item{req_root}{The root URL of the request to Tableau server.}
}
\value{
The root URL where requests to Tableau server should be directed,
made of the URL of the dashboard plus a unique session ID.
}
\description{
Get Tableau session ID
}
