% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableau.R
\name{tableau_tooltip}
\alias{tableau_tooltip}
\title{Make a request to Tableau tooltip server}
\usage{
tableau_tooltip(url, req_root, post)
}
\arguments{
\item{url}{The URL of the Tableau dashboard.}

\item{req_root}{The root URL of the request to Tableau server.}

\item{post}{Data to POST identifying the tooltip requested.}
}
\value{
HTML tooltip text.
}
\description{
The data identifying the specific tooltip to request should have `--[boundary]`
in place of the initial boundary value so that a random boundary value may
be substituted where appropriate.
}
