% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdriver.R
\name{webdriver}
\alias{webdriver}
\alias{webdriver_get}
\alias{webdriver_open}
\alias{webdriver_wait_for_element}
\alias{webdriver_close}
\alias{webdriver_commands}
\title{Functions for RSelenium}
\usage{
webdriver_get(
  uuid,
  webdriver_options = list(chromever = "latest", check = TRUE, verbose = FALSE)
)

webdriver_open(url, chromever = "latest", check = TRUE, verbose = FALSE)

webdriver_wait_for_element(webdriver, By, value, timeout)

webdriver_close(webdriver)

webdriver_commands(webdriver, uuid)
}
\arguments{
\item{uuid}{The UUID of the dataset from datasets.json.}

\item{webdriver_options}{Optionally, arguments for \code{webdriver_open} provided
as a named list.}

\item{url}{The URL to navigate to.}

\item{chromever}{See parameter in \code{\link[RSelenium]{rsDriver}}.}

\item{check}{See parameter in \code{\link[RSelenium]{rsDriver}}.}

\item{verbose}{See parameter in \code{\link[RSelenium]{rsDriver}}.}

\item{webdriver}{Selenium server object.}

\item{By}{Type of selector to use.}

\item{value}{Value for selector.}

\item{timeout}{Timeout in seconds.}
}
\value{
The function \code{webdriver_open} returns an HTML object created by \code{\link[xml2]{read_html}}.
}
\description{
Navigate to a webpage using RSelenium server and tidy up afterwards.
This is used for pages requiring JavaScript to load correctly.
The primary command is \code{webdriver_get}, which is called by
\code{\link[Covid19CanadaData]{dl_dataset}} when JavaScript is required. Note
that this function requires a recent version of the Chrome/Chromium
browser. If the latest version of Chrome/Chromium is not used, the version of
Chromedriver must be manually specified in \code{webdriver_options$chromever}.
Run \code{binman::list_versions("chromedriver")} for available versions.
}
