% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_archive.R
\name{dl_archive}
\alias{dl_archive}
\title{Download archived versions of a dataset catalogued in Covid19CanadaArchive}
\usage{
dl_archive(
  uuid,
  date,
  after,
  before,
  sep,
  sheet,
  remove_duplicates = TRUE,
  overwrite = FALSE,
  path = NULL
)
}
\arguments{
\item{uuid}{The UUID of the dataset from datasets.json.}

\item{date}{A character string in YYYY-MM-DD format specifying the date of
data to return. Also accepts "latest" (the default), "first" and "all".}

\item{after}{A character string in YYYY-MM-DD format specifying that data
from this date or later should be returned. Ignored if `date` is defined.}

\item{before}{A character string in YYYY-MM-DD format specifying that data
from this date or earlier should be returned. Ignored if `date` is defined.}

\item{sep}{Optional. The separator to use when reading CSV files.
Defaults to ",".}

\item{sheet}{Optional. An integer or name specifying the sheet to return for
an XLSX or XLS file (by default, reads sheet 1 with a warning).}

\item{remove_duplicates}{Remove duplicate files from the sample after date
filtering? Defaults to `TRUE`. Ignored if a single date is requested.}

\item{overwrite}{When saving to disk (by specifying the argument `path`),
should existing files be overwritten? If `FALSE` (the default), only new
files will be downloaded.}

\item{path}{A character string specifying the file path to write the
specified dataset as a file/files. This value is NULL by default, resulting
in the dataset being returned as a list.}
}
\value{
A list containing the specified dataset(s) as R objects (the default)
or written to a file/files by (by specifying the argument `path`). In the
latter case, a vector of file paths is also returned invisibly.
}
\description{
Download the archived version of a dataset listed in datasets.json of
Covid19CanadaArchive (https://github.com/ccodwg/Covid19CanadaArchive/blob/master/datasets.json).
Data can either be imported into R (the default) or written to a file by
specifying the `file` argument. Currently, only CSV, JSON, XLSX, XLS, image
and HTML datasets are supported for reading into R. If no date filtering is
specified, the latest file is returned by default.
}
