% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_dataset.R
\name{dl_dataset}
\alias{dl_dataset}
\title{Download current version of a dataset catalogued in Covid19CanadaArchive}
\usage{
dl_dataset(
  uuid,
  file = NULL,
  sep = NULL,
  sheet = NULL,
  webdriver_options = list(chromever = "latest", check = TRUE, verbose = FALSE)
)
}
\arguments{
\item{uuid}{The UUID of the dataset from datasets.json.}

\item{file}{A character string specifying the location to write the specified
dataset as a file (NULL by default, resulting in the dataset being returned
as a data frame).}

\item{sep}{The separator to use when reading CSV files. Defaults to ",".}

\item{sheet}{An integer specifying the sheet to return for an XLSX or XLS
file (by default, reads sheet 1 with a warning).}

\item{webdriver_options}{Optionally, arguments for \code{\link[Covid19CanadaData]{webdriver_open}}
provided as a named list.}
}
\value{
The specified dataset either as a dataset loaded in R (the default)
or written to a file by (by specifying the argument `file`).
}
\description{
Download the current version of an active dataset listed in datasets.json of
Covid19CanadaArchive (https://github.com/ccodwg/Covid19CanadaArchive/blob/master/datasets.json).
Data can either be imported into R (the default) or written to a file by
specifying the `file` argument. Currently, only CSV, JSON, XLSX, XLS, image
and HTML datasets are supported for reading into R.
}
\details{
Retrieving the HTML of webpages requiring JavaScript (args: js = True in
datasets.json) uses \code{\link[RSelenium]{rsDriver}} from the
\code{RSelenium} package. This requires a recent version of the Chrome/Chromium
browser to be installed. If you are using a recent version but not the latest
version, a version number for Chromedriver must be provided to
\code{webdriver_options}. See \code{\link[Covid19CanadaData]{webdriver_get}}
for more details.
}
\examples{
\dontrun{
# get PHAC epidemiology update CSV
dl_dataset("f7db31d0-6504-4a55-86f7-608664517bdb")

# get Saskatchewan total cases CSV
dl_dataset("61cfdd06-7749-4ae6-9975-d8b4f10d5651")
}
}
