% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_dataset_helpers}
\alias{process_dataset_helpers}
\alias{helper_cum_current}
\title{Helper functions for process_dataset}
\usage{
helper_cum_current(.data, loc = c("prov", "hr"), val, prov, date_current)
}
\description{
Helper functions for \code{\link{process_dataset}}
}
