% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdriver.R
\name{webdriver}
\alias{webdriver}
\alias{webdriver_open}
\alias{webdriver_close}
\title{Functions for RSelenium}
\usage{
webdriver_open(url, headless = TRUE, check = FALSE, verbose = FALSE)

webdriver_close(webdriver)
}
\arguments{
\item{url}{The URL to navigate to.}

\item{headless}{Run in headless mode? Default: TRUE. Non-headless mode can
be useful for debugging.}

\item{check}{See parameter in \code{\link[RSelenium]{rsDriver}}.}

\item{verbose}{See parameter in \code{\link[RSelenium]{rsDriver}}.}

\item{webdriver}{Selenium server object.}
}
\description{
Navigate to a webpage using RSelenium server and tidy up afterwards.
}
